webpki: move crl_error and pki_error tests
This commit moves the unit tests for the `crl_error` and `pki_error`
helpers to be defined alongside the helpers themselves, in `mod.rs`, as
opposed to `verify.rs`.
