+++ rustls/src/webpki/mod.rs
@@ -66,3 +66,84 @@ fn crl_error(e: webpki::Error) -> CertRevocationListError {
         _ => CertRevocationListError::Other(Arc::new(e)),
     }
 }
+
+mod tests {
+    #[test]
+    fn pki_crl_errors() {
+        use super::{pki_error, CertRevocationListError, CertificateError, Error};
+
+        // CRL signature errors should be turned into BadSignature.
+        assert_eq!(
+            pki_error(webpki::Error::InvalidCrlSignatureForPublicKey),
+            Error::InvalidCertRevocationList(CertRevocationListError::BadSignature),
+        );
+        assert_eq!(
+            pki_error(webpki::Error::UnsupportedCrlSignatureAlgorithm),
+            Error::InvalidCertRevocationList(CertRevocationListError::BadSignature),
+        );
+        assert_eq!(
+            pki_error(webpki::Error::UnsupportedCrlSignatureAlgorithmForPublicKey),
+            Error::InvalidCertRevocationList(CertRevocationListError::BadSignature),
+        );
+
+        // Revoked cert errors should be turned into Revoked.
+        assert_eq!(
+            pki_error(webpki::Error::CertRevoked),
+            Error::InvalidCertificate(CertificateError::Revoked),
+        );
+
+        // Issuer not CRL signer errors should be turned into IssuerInvalidForCrl
+        assert_eq!(
+            pki_error(webpki::Error::IssuerNotCrlSigner),
+            Error::InvalidCertRevocationList(CertRevocationListError::IssuerInvalidForCrl)
+        );
+    }
+
+    #[test]
+    fn crl_error_from_webpki() {
+        use super::{crl_error, CertRevocationListError::*};
+
+        let testcases = &[
+            (webpki::Error::InvalidCrlSignatureForPublicKey, BadSignature),
+            (
+                webpki::Error::UnsupportedCrlSignatureAlgorithm,
+                BadSignature,
+            ),
+            (
+                webpki::Error::UnsupportedCrlSignatureAlgorithmForPublicKey,
+                BadSignature,
+            ),
+            (webpki::Error::InvalidCrlNumber, InvalidCrlNumber),
+            (
+                webpki::Error::InvalidSerialNumber,
+                InvalidRevokedCertSerialNumber,
+            ),
+            (webpki::Error::IssuerNotCrlSigner, IssuerInvalidForCrl),
+            (webpki::Error::MalformedExtensions, ParseError),
+            (webpki::Error::BadDer, ParseError),
+            (webpki::Error::BadDerTime, ParseError),
+            (
+                webpki::Error::UnsupportedCriticalExtension,
+                UnsupportedCriticalExtension,
+            ),
+            (webpki::Error::UnsupportedCrlVersion, UnsupportedCrlVersion),
+            (webpki::Error::UnsupportedDeltaCrl, UnsupportedDeltaCrl),
+            (
+                webpki::Error::UnsupportedIndirectCrl,
+                UnsupportedIndirectCrl,
+            ),
+            (
+                webpki::Error::UnsupportedRevocationReason,
+                UnsupportedRevocationReason,
+            ),
+        ];
+        for t in testcases {
+            assert_eq!(crl_error(t.0), t.1);
+        }
+
+        assert!(matches!(
+            crl_error(webpki::Error::NameConstraintViolation),
+            Other(_)
+        ));
+    }
+}
