+++ rustls/src/webpki/verify.rs
@@ -377,85 +377,7 @@ pub(crate) fn verify_tls13(
 
 #[cfg(test)]
 mod tests {
-    use super::super::crl_error;
     use super::*;
-    use crate::error::CertRevocationListError;
-
-    #[test]
-    fn pki_crl_errors() {
-        // CRL signature errors should be turned into BadSignature.
-        assert_eq!(
-            pki_error(webpki::Error::InvalidCrlSignatureForPublicKey),
-            Error::InvalidCertRevocationList(CertRevocationListError::BadSignature),
-        );
-        assert_eq!(
-            pki_error(webpki::Error::UnsupportedCrlSignatureAlgorithm),
-            Error::InvalidCertRevocationList(CertRevocationListError::BadSignature),
-        );
-        assert_eq!(
-            pki_error(webpki::Error::UnsupportedCrlSignatureAlgorithmForPublicKey),
-            Error::InvalidCertRevocationList(CertRevocationListError::BadSignature),
-        );
-
-        // Revoked cert errors should be turned into Revoked.
-        assert_eq!(
-            pki_error(webpki::Error::CertRevoked),
-            Error::InvalidCertificate(CertificateError::Revoked),
-        );
-
-        // Issuer not CRL signer errors should be turned into IssuerInvalidForCrl
-        assert_eq!(
-            pki_error(webpki::Error::IssuerNotCrlSigner),
-            Error::InvalidCertRevocationList(CertRevocationListError::IssuerInvalidForCrl)
-        );
-    }
-
-    #[test]
-    fn crl_error_from_webpki() {
-        use crate::CertRevocationListError::*;
-        let testcases = &[
-            (webpki::Error::InvalidCrlSignatureForPublicKey, BadSignature),
-            (
-                webpki::Error::UnsupportedCrlSignatureAlgorithm,
-                BadSignature,
-            ),
-            (
-                webpki::Error::UnsupportedCrlSignatureAlgorithmForPublicKey,
-                BadSignature,
-            ),
-            (webpki::Error::InvalidCrlNumber, InvalidCrlNumber),
-            (
-                webpki::Error::InvalidSerialNumber,
-                InvalidRevokedCertSerialNumber,
-            ),
-            (webpki::Error::IssuerNotCrlSigner, IssuerInvalidForCrl),
-            (webpki::Error::MalformedExtensions, ParseError),
-            (webpki::Error::BadDer, ParseError),
-            (webpki::Error::BadDerTime, ParseError),
-            (
-                webpki::Error::UnsupportedCriticalExtension,
-                UnsupportedCriticalExtension,
-            ),
-            (webpki::Error::UnsupportedCrlVersion, UnsupportedCrlVersion),
-            (webpki::Error::UnsupportedDeltaCrl, UnsupportedDeltaCrl),
-            (
-                webpki::Error::UnsupportedIndirectCrl,
-                UnsupportedIndirectCrl,
-            ),
-            (
-                webpki::Error::UnsupportedRevocationReason,
-                UnsupportedRevocationReason,
-            ),
-        ];
-        for t in testcases {
-            assert_eq!(crl_error(t.0), t.1);
-        }
-
-        assert!(matches!(
-            crl_error(webpki::Error::NameConstraintViolation),
-            Other(_)
-        ));
-    }
 
     #[test]
     fn certificate_debug() {
